+++ common/scripts/definitions/utils.ts
@@ -32,6 +32,40 @@ export interface BulletDef{
     criticalMult?:number
     on_hit_explosion?:string
 }
+export type ItemQualitySetting={
+    color1:string
+    color2:string
+}
+export const ItemQualitySettings:Record<ItemQuality,ItemQualitySetting>={
+    [ItemQuality.Common]:{
+        color1:"#eeeeee",
+        color2:"#a0a0a0"
+    },
+    [ItemQuality.Uncommon]:{
+        color1:"#11ef45",
+        color2:"#0c913a",
+    },
+    [ItemQuality.Rare]:{
+        color1:"#3533ee",
+        color2:"#15118a"
+    },
+    [ItemQuality.Epic]:{
+        color1:"#9309de",
+        color2:"#3b0b7d"
+    },
+    [ItemQuality.Mythic]:{
+        color1:"#f0d107",
+        color2:"#ab8c0f"
+    },
+    [ItemQuality.Legendary]:{
+        color1:"#ed092c",
+        color2:"#a3050a"
+    },
+    [ItemQuality.Developer]:{
+        color1:"#eeeeee",
+        color2:"#eeeeee"
+    },
+}
 export enum InventoryItemType{
     gun,
     ammo,